Multinomial Regression using R shiny for child health related risks in England.
This analysis was aim to make use of the publicly available data to check whether there is a relationship between the Local Authority Health Profiles and Local indicators of deprivation in terms of child health-related risk factors using Multinomial regression in R shiny. The insights and knowledge derived will be a guide to the local authorities to understand the health-related risk factors of child in England and formulate new guidelines or policies.